+Clean unloading, part II: the popviewer extension.
Define a proper minor mode and an unload function to be able to
cleanly toggle/disable the functionality.

Define two new functions (`sr-ad-enable' and `sr-ad-disable') to
enable/disable certain (pieces of) advice and update the relevant
function definitions at the same time.

[Using advice is evil and I'd like to avoid it as much as possible,
but it's certainly better than hard-redefining the functions.]

Name the viewer frame "Sunrise Viewer" instead of "Sunrise Viewer
Frame". It's shorter and should make more sense even to users not used
to Emacs jargon.

The reason for not putting contents of `sr-popviewer-setup-windows'
directly into the defadvice is that the byte-compiler apparently
doesn't compile advice definitions (and the `compile' flag is
apparently only to be used together with `activate', which we don't
want here), so the unexpanded `case' would trigger a runtime error
without `cl-macs' loaded.